feat: write logs to a user directory
Write the log file to a user-specific directory, depending on the OS:
- `~/Library/Logs/DAISY Ace/ace.log` on macOS
- `~\AppData\Local\DAISY Ace\Log\ace.log`  on Windows
- `~/.local/state/DAISY Ace/ace.log` on Linux

Also augment Winston with a `logAndExit` method to propertly flush
the logs to files before exitting (see winstonjs/winston#228).

In the CLI, move the main processing to an `async` function and
wait for a timeout when logging and exitting.

Closes #50.