fix: patch aXe to support namespaced element names
aXe is creating CSS selectors for relevant elements in the checked
content. When faced with XML-namespaced element names (e.g. MathML
elements), it used the prefixed element name as a selector for the
element, which caused a syntax error when matching against this
selector.

To fix this, we copy part of aXe's code in our own file as a monkey
patch.
This must be removed when dequelabs/axe-core#563 is fixed.

Closes #57