perf: run HTML checks concurrently
Runs all Content Document checks concurrently, earch within a new
page of a shared Chrome browser instance.

The results are aggregated in document order.

The concurrency level is currently hard-coded to 4, but it should
be made configurable when implementing issue #77.

Based on very naïve –and probably not very significant– benchmarking,
the execution time is reduced by a factor of at least 2 on large
publications.