fix(report): report looks ok even if javascript is disabled
All violations are listed in the raw HTML, and if you have JS, you can use filtering and sorting. This results in some duplication of code for populating the table (happens in the report generation as well as in the report's own javascript).

The tabs and a few other stylistic aspects are disabled unless js is present.

Closes #119