feat(cli): optionally return error code 2 when violations are found
The CLI can now return the error code `2` if there was any validation
errors/warnings.

The report json object is returned by `ace()`, and the calling function
in the CLI can see the results of the validation.

Also output if the package doc had any issues (same as we already do for
content docs).

Also started adding a test, and a test EPUB with known violation, to
trigger return code 2. Pending #150 before it will work, so it's disabled
until then.

Closes #113.