fix(report): remove @path property from iframes and images
The json data is scrubbed clean of @path properties after the images are
copied. If we aren't storing the report on disk, the json data is scrubbed
before it gets written to the console.

I'd rather separate out any data that needs to be copied and not have it
stored in the report in the first place, but that would involve some
redesign. Maybe a good idea for the future.

fs-extra's `writeFile` was causing concurrency issues, so it's been
replaced with `writeFileSync`.

Closes #64